+++ src/main/java/lotto/view/dto/LottoWinningNumberCreateRequest.java
@@ -1,4 +1,4 @@
-package lotto.controller.dto;
+package lotto.view.dto;
 
 import java.util.ArrayList;
 import java.util.List;
