+++ src/test/java/lotto/dto/LottoBonusNumberCreateRequestTest.java
@@ -1,6 +1,6 @@
 package lotto.dto;
 
-import lotto.controller.dto.LottoBonusNumberCreateRequest;
+import lotto.view.dto.LottoBonusNumberCreateRequest;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
